requirements.txt
googletrans~=4.0.0rc1
google-cloud-translate~=3.8.2
python-dotenv~=0.21.1
